fix(SFINT-2500): Add ts-es5-istanbul-coverage loader
This package will add the istanbul ignore statement where they need to be to circumvent the issue with TypeScript transpilation to ES5.
This will only apply when the code is transpiled for unit tests, not for production.
Also remove all the istanbul ignore comments now that they're not needed anymore.